Move lwrp to real lwrp
All resources and providers were defined by subclassing lwrpbase
classes. The lwrp dsl make it simpler.
With this change, the cookbook is compatible with chef 11 again
